libtock: tock.c: be u32/u64 compatible
memop, allow, and subscribe use register-with types rather than int

+++ libtock/tock.h
@@ -124,7 +124,7 @@ typedef struct {
   statuscode_t status;
   // Optional return data depending on the memop variant called. Only set if
   // status is `TOCK_STATUSCODE_SUCCESS`.
-  uint32_t data;
+  uintptr_t data;
 } memop_return_t;
 
 // Return structure for a Yield-WaitFor syscall. The return value are the
@@ -216,7 +216,7 @@ __attribute__ ((warn_unused_result))
 allow_ro_return_t allow_readonly(uint32_t driver, uint32_t allow, const void* ptr, size_t size);
 
 // Call the memop syscall.
-memop_return_t memop(uint32_t op_type, int arg1);
+memop_return_t memop(uint32_t op_type, uintptr_t arg1);
 
 // Wrappers around memop to support app introspection
 void* tock_app_memory_begins_at(void);
